preliminary error component and client side page not found routing
