+++ src/components/FallbackError/FallbackError.tsx
@@ -1,0 +1,38 @@
+import React, { Component, ErrorInfo, ReactNode } from "react";
+
+interface Props {
+  children?: ReactNode;
+}
+
+interface State {
+  hasError: boolean;
+}
+
+class FallbackError extends Component<Props, State> {
+  public state: State = {
+    hasError: false,
+  };
+
+  public static getDerivedStateFromError(_: Error): State {
+    // Update state so the next render will show the fallback UI.
+    return { hasError: true };
+  }
+
+  public componentDidCatch(error: Error, errorInfo: ErrorInfo) {
+    console.error("Uncaught error:", error, errorInfo);
+  }
+
+  public render() {
+    if (this.state.hasError) {
+      return (
+        <div className="bg-dark flex-fill">
+          <h1 className="text-light">Sorry.. we experienced an error</h1>
+        </div>
+      );
+    }
+
+    return this.props.children;
+  }
+}
+
+export default FallbackError;
