+++ src/components/FallbackError/index.ts
@@ -1,0 +1,3 @@
+import FallbackError from "./FallbackError";
+
+export default FallbackError;
